Inline menu sheet into marking bottom sheet
