--- apps/frontend/app/components/MenuSheet/types.ts
@@ -1,3 +1,0 @@
-export interface MenuSheetProps {
-	closeSheet: () => void;
-}
